+++ changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [5.7.1 (unreleased)](https://github.com/kdeldycke/workflows/compare/v5.7.0...main)
+
+> [!IMPORTANT]
+> This version is not released yet and is under active development.
+
 ## [5.7.0 (2026-02-09)](https://github.com/kdeldycke/workflows/compare/v5.6.2...v5.7.0)
 
 - Add `gha-utils sphinx-linkcheck` command to detect broken auto-generated links.